Fixed server messages
Yes, I'm too lazy to change branches just for this.

+++ src/ChannelServer/Network/ChannelPacketSender.cs
@@ -875,7 +875,7 @@ namespace Melia.Channel.Network
 
 			packet.PutInt(clientMessage);
 			packet.PutByte((byte)parameters.Length);
-			packet.PutShort(1); // type? 0 = also show in red letters on the screen
+			packet.PutByte(1); // type? 0 = also show in red letters on the screen
 			packet.PutLong(0); // added i219527
 			packet.PutByte(0); // added i336041
 			packet.PutByte(0); // added i339415
